♻️ Refactor and streamline the QSCI example and make it Qiskit 2.0 compatible.
