+++ examples/qsci_h2.py
@@ -19,125 +19,46 @@
 # /// script
 # requires-python = ">=3.10,<3.14"
 # dependencies = [
-#   "iqm-qdmi",
-#   "mqt-core[qiskit]~=3.6.0",
-#   "qiskit>=1.1,<2.0",
-#   "qiskit-algorithms>=0.3.0,<0.4.0",
-#   "qiskit-nature>=0.7.2",
-#   "pyscf>=2.11.0; platform_system != 'Windows'",
+#   "iqm-qdmi[qiskit]",
+#   "qiskit-nature[pyscf]>=0.7.2",
 # ]
+#
+# [tool.uv.sources]
+# iqm-qdmi = { path = ".." }
 # ///
 
-"""Run an H2 QSCI example through the IQM QDMI Qiskit backend."""
+"""Run QSCI on the H2 molecule through the IQM QDMI Qiskit backend."""
 
 from __future__ import annotations
 
 import argparse
-import os
-import platform
 import sys
-from pathlib import Path
-from typing import TYPE_CHECKING
+from functools import reduce
+from typing import TYPE_CHECKING, cast
 
-if __package__ in {None, ""}:
-    sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-
-from examples import qsci_postprocess
+import numpy as np
+import scipy.linalg as sla
+from mqt.core.plugins.qiskit.estimator import QDMIEstimator
 from mqt.core.plugins.qiskit.provider import QDMIProvider
 from mqt.core.plugins.qiskit.sampler import QDMISampler
-from qiskit import transpile
-from qiskit.primitives import BackendEstimator  # ty: ignore[unresolved-import]
-from qiskit.quantum_info import SparsePauliOp
-from qiskit_algorithms import VQE, NumPyMinimumEigensolver
-from qiskit_algorithms.optimizers import L_BFGS_B
-from qiskit_nature.second_q.algorithms import GroundStateEigensolver
-from qiskit_nature.second_q.algorithms.initial_points import MP2InitialPoint
+from pyscf import ao2mo, gto, scf
+from qiskit.compiler import transpile
+from qiskit_algorithms.minimum_eigensolvers.vqe import VQE
+from qiskit_algorithms.optimizers.l_bfgs_b import L_BFGS_B
 from qiskit_nature.second_q.circuit.library import UCCSD, HartreeFock
 from qiskit_nature.second_q.drivers import PySCFDriver
 from qiskit_nature.second_q.mappers import JordanWignerMapper
-from qiskit_nature.second_q.operators import FermionicOp
+from scipy.sparse import csr_matrix
+from scipy.sparse.linalg import eigs
 
 from iqm.qdmi.qiskit import IQMBackend
 
 if TYPE_CHECKING:
-    from mqt.core.plugins.qiskit.backend import QDMIBackend
-    from qiskit.circuit import QuantumCircuit
+    from qiskit.quantum_info import SparsePauliOp
 
 ATOM = "H 0 0 0; H 0 0 1;"
 BASIS = "sto-3g"
-
-
-def parse_args() -> argparse.Namespace:
-    """Parse CLI arguments for the QSCI example.
-
-    Returns:
-        Parsed command-line arguments.
-    """
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("--backend", choices=("iqm", "sim"), default="iqm")
-    parser.add_argument("--shots", type=int, default=2048)
-    parser.add_argument("--maxiter", type=int, default=30)
-    parser.add_argument("--cutoff", type=int, default=10)
-    return parser.parse_args()
-
-
-def make_backend(backend_kind: str) -> QDMIBackend:
-    """Create the selected QDMI backend for the example run.
-
-    Args:
-        backend_kind: Backend mode selected on the command line.
-
-    Returns:
-        The requested QDMI backend.
-
-    Raises:
-        SystemExit: If required IQM credentials are missing or no simulator backend is available.
-    """
-    if backend_kind == "iqm":
-        token = os.getenv("IQM_TOKEN") or os.getenv("RESONANCE_API_KEY")
-        tokens_file = os.getenv("IQM_TOKENS_FILE")
-        if token is None and tokens_file is None:
-            msg = "Set RESONANCE_API_KEY, IQM_TOKEN, or IQM_TOKENS_FILE to run this example on IQM hardware."
-            raise SystemExit(msg)
-
-        return IQMBackend(
-            base_url=os.getenv("IQM_BASE_URL"),
-            token=token,
-            tokens_file=tokens_file,
-            qc_id=os.getenv("IQM_QC_ID"),
-            qc_alias=os.getenv("IQM_QC_ALIAS"),
-        )
-
-    provider = QDMIProvider()
-    try:
-        return provider.get_backend("MQT Core DDSIM QDMI Device")
-    except ValueError:
-        simulator_backends = provider.backends(name="DDSIM")
-        if len(simulator_backends) == 1:
-            return simulator_backends[0]
-        if not simulator_backends:
-            msg = "No QDMI simulator backend is available through QDMIProvider()."
-            raise SystemExit(msg) from None
-
-        backend_names = ", ".join(backend.name or "<unnamed>" for backend in simulator_backends)
-        msg = f"Multiple simulator backends matched this example: {backend_names}."
-        raise SystemExit(msg) from None
-
-
-def transpile_for_backend(circuit: QuantumCircuit, backend: QDMIBackend, backend_kind: str) -> QuantumCircuit:
-    """Transpile a circuit for the selected backend kind.
-
-    Args:
-        circuit: Circuit to transpile.
-        backend: Backend used for backend-aware transpilation.
-        backend_kind: Backend mode selected on the command line.
-
-    Returns:
-        The transpiled circuit.
-    """
-    if backend_kind == "sim":
-        return transpile(circuit, basis_gates=["u", "cx", "measure"], optimization_level=1)
-    return transpile(circuit, backend=backend, optimization_level=3)
+EXPECTED_ENERGY = -1.101150
 
 
 def main() -> None:
@@ -147,25 +68,29 @@ def main() -> None:
         SystemExit: If backend setup fails, required chemistry data is missing,
             or the backend returns an unexpected shot count.
     """
-    if platform.system() == "Windows":
+    if sys.platform == "win32":
         msg = "QSCI requires PySCF, which is not supported on Windows."
         raise SystemExit(msg)
 
-    args = parse_args()
-    backend = make_backend(args.backend)
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--backend", choices=("iqm", "sim"), default="iqm")
+    parser.add_argument("--shots", type=int, default=8192)
+    parser.add_argument("--maxiter", type=int, default=30)
+    parser.add_argument("--cutoff", type=int, default=10)
+    args = parser.parse_args()
+
+    if args.backend == "iqm":
+        backend = IQMBackend()
+    else:
+        provider = QDMIProvider()
+        try:
+            backend = provider.get_backend("MQT Core DDSIM QDMI Device")
+        except ValueError:
+            msg = "MQT Core DDSIM QDMI Device backend is not available through QDMIProvider()."
+            raise SystemExit(msg) from None
 
     driver = PySCFDriver(atom=ATOM, basis=BASIS)
     problem = driver.run()
-    if problem.num_spatial_orbitals is None:
-        msg = "num_spatial_orbitals is required"
-        raise SystemExit(msg)
-    if problem.num_particles is None:
-        msg = "num_particles is required"
-        raise SystemExit(msg)
-    if problem.nuclear_repulsion_energy is None:
-        msg = "nuclear_repulsion_energy is required"
-        raise SystemExit(msg)
-
     mapper = JordanWignerMapper()
     initial_state = HartreeFock(problem.num_spatial_orbitals, problem.num_particles, mapper)
     ansatz = UCCSD(
@@ -174,70 +99,195 @@ def main() -> None:
         mapper,
         initial_state=initial_state,
     )
-    second_q_operator = problem.second_q_ops()[0]
-    if not isinstance(second_q_operator, FermionicOp):
-        msg = "The QSCI example requires a fermionic operator."
-        raise SystemExit(msg)
-    observable = mapper.map(second_q_operator)
-    if not isinstance(observable, SparsePauliOp):
-        msg = "Mapped observable must be a SparsePauliOp."
-        raise SystemExit(msg)
+
     if backend.num_qubits < ansatz.num_qubits:
         msg = f"Selected backend exposes {backend.num_qubits} qubits, but the QSCI example needs {ansatz.num_qubits}."
         raise SystemExit(msg)
 
-    vqe_ansatz = transpile_for_backend(ansatz, backend, args.backend) if args.backend == "sim" else ansatz
-    mp2_initial_point = MP2InitialPoint()
-    mp2_initial_point.compute(ansatz=ansatz, problem=problem)
-    # qiskit_algorithms.VQE still expects the Estimator V1 interface.
-    estimator = BackendEstimator(
-        backend=backend,
-        options={"shots": args.shots},
-        skip_transpilation=args.backend == "sim",
-    )
+    ansatz = transpile(ansatz, backend=backend, optimization_level=3)
+    estimator = QDMIEstimator(backend, options={"default_shots": args.shots})
     vqe = VQE(
         estimator,
-        vqe_ansatz,
+        ansatz,
         L_BFGS_B(maxiter=args.maxiter),
-        initial_point=mp2_initial_point.to_numpy_array(),
     )
+
+    hamiltonian = problem.hamiltonian
+    second_q_operator = hamiltonian.second_q_op()
+    observable = cast("SparsePauliOp", mapper.map(second_q_operator))
     result = vqe.compute_minimum_eigenvalue(operator=observable)
-    if result.optimal_parameters is None:
+    optimal_parameters = result.optimal_parameters
+    if optimal_parameters is None:
         msg = "VQE must return optimal parameters."
         raise SystemExit(msg)
 
-    trained_ansatz = vqe_ansatz.assign_parameters(result.optimal_parameters)
-    if trained_ansatz is None:
-        msg = "assign_parameters must return a bound circuit."
-        raise SystemExit(msg)
-
-    sample_circuit = trained_ansatz.copy()
-    sample_circuit.measure_all()
-    transpiled_sample = transpile_for_backend(sample_circuit, backend, args.backend)
+    ansatz.assign_parameters(optimal_parameters, inplace=True)
+    ansatz.measure_active()
     sampler = QDMISampler(backend, default_shots=args.shots)
-    job = sampler.run([(transpiled_sample,)])
-    data = job.result()[0].data
-    counts = data["meas"].get_counts() if "meas" in data else next(iter(data.values())).get_counts()
-    if sum(counts.values()) != args.shots:
-        msg = f"Expected {args.shots} shots, but observed {sum(counts.values())}."
+    job = sampler.run([(ansatz,)])
+    counts = job.result()[0].data["meas"].get_counts()
+    performed_shots: int = sum(counts.values())
+    if performed_shots != args.shots:
+        msg = f"Expected {args.shots} shots, but observed {performed_shots}."
         raise SystemExit(msg)
 
-    eigval = qsci_postprocess.postprocess_counts(ATOM, BASIS, counts, cutoff=args.cutoff)
-    qsci_energy = eigval + problem.nuclear_repulsion_energy
-
-    exact_solver = GroundStateEigensolver(mapper, NumPyMinimumEigensolver())
-    exact_result = exact_solver.solve(problem)
-    exact_energies = getattr(exact_result, "total_energies", None)
-    if exact_energies is None:
-        msg = "The exact solver must return total energies."
-        raise SystemExit(msg)
-    exact_energy = float(exact_energies[0])
-    energy_difference = abs(exact_energy - qsci_energy)
+    eigval = _postprocess_counts(ATOM, BASIS, counts, cutoff=args.cutoff)
+    qsci_energy = eigval + (problem.hamiltonian.nuclear_repulsion_energy or 0.0)
 
     print(f"Sampled {args.shots} shots from the optimized ansatz.")
     print(f"QSCI total energy: {qsci_energy:.6f}")
-    print(f"Exact total energy: {exact_energy:.6f}")
-    print(f"Absolute energy difference: {energy_difference:.6f}")
+    energy_difference = abs(EXPECTED_ENERGY - qsci_energy)
+    print(f"Absolute energy difference: {energy_difference:.6f} to the expected energy of {EXPECTED_ENERGY:.6f}")
+
+
+def _full_mbh_canonical_from_qsci(
+    basis: list[int],
+    h_diag: np.ndarray,
+    h_off_diag: np.ndarray,
+    u_full: np.ndarray,
+    num_qubits: int,
+    tol: float = 1e-4,
+) -> tuple[list[int], list[int], list[float]]:
+    """Calculate a sparse Hamiltonian representation for the selected QSCI basis.
+
+    Args:
+        basis: Bitstring basis states kept for the reduced Hamiltonian.
+        h_diag: Diagonal one-body contributions in the spin-orbital basis.
+        h_off_diag: Off-diagonal one-body contributions in the spin-orbital basis.
+        u_full: Two-body interaction tensor in chemist notation.
+        num_qubits: Number of spin orbitals in the encoded problem.
+        tol: Numerical threshold used to discard negligible terms.
+
+    Returns:
+        Sparse matrix triplets as row indices, column indices, and data values.
+    """
+    row, col, data = [], [], []
+
+    for index, state in enumerate(basis):
+        occupations = [int(digit) for digit in f"{state:0{num_qubits}b}"]
+        occupations_up = occupations[: num_qubits // 2]
+        occupations_down = occupations[num_qubits // 2 :]
+
+        diagonal_interaction = sum(
+            u_full[orbital, orbital, partner, partner] * occupations_up[orbital] * occupations_down[partner]
+            for orbital in range(num_qubits // 2)
+            for partner in range(num_qubits // 2)
+        )
+        diagonal_value = sum(occupations * h_diag) + diagonal_interaction
+        if abs(diagonal_value) > tol:
+            row.append(index)
+            col.append(index)
+            data.append(float(diagonal_value))
+
+        for column_index in range(index + 1, len(basis)):
+            differing_bits = f"{state ^ basis[column_index]:0{num_qubits}b}"
+            bit_indices = [offset for offset, bit in enumerate(differing_bits) if bit == "1"]
+
+            if len(bit_indices) == 2:
+                first, second = bit_indices
+                if first < num_qubits // 2 and second < num_qubits // 2:
+                    off_diagonal_value = h_off_diag[first, second] + sum(
+                        u_full[first, second, partner, partner] * occupations[partner + num_qubits // 2]
+                        for partner in range(num_qubits // 2)
+                    )
+                elif first >= num_qubits // 2 and second >= num_qubits // 2:
+                    off_diagonal_value = h_off_diag[first - num_qubits // 2, second - num_qubits // 2] + sum(
+                        u_full[partner, partner, first - num_qubits // 2, second - num_qubits // 2]
+                        * occupations[partner]
+                        for partner in range(num_qubits // 2)
+                    )
+                else:
+                    continue
+                if abs(off_diagonal_value) > tol:
+                    sign = (-1) ** sum(occupations[min(first, second) + 1 : max(first, second)])
+                    row.extend((index, column_index))
+                    col.extend((column_index, index))
+                    data.extend((float(sign * off_diagonal_value), float(sign * off_diagonal_value)))
+
+            elif len(bit_indices) == 4:
+                first, second, third, fourth = bit_indices
+                if first < num_qubits // 2 <= third and second < num_qubits // 2 <= fourth:
+                    off_diagonal_value = u_full[first, second, third - num_qubits // 2, fourth - num_qubits // 2]
+                    if abs(off_diagonal_value) > tol:
+                        sign_up = (-1) ** sum(occupations[first + 1 : second])
+                        sign_down = (-1) ** sum(occupations[third + 1 : fourth])
+                        sign = sign_up * sign_down
+                        row.extend((index, column_index))
+                        col.extend((column_index, index))
+                        data.extend((float(sign * off_diagonal_value), float(sign * off_diagonal_value)))
+
+    return row, col, data
+
+
+def _postprocess_counts(atom: str, basis: str, counts: dict[str, int], *, cutoff: int = 10) -> float:
+    """Postprocess sampled bitstrings into a QSCI energy estimate.
+
+    Args:
+        atom: Molecular geometry specification passed to PySCF.
+        basis: Basis-set label passed to PySCF.
+        counts: Sampled computational-basis bitstring counts.
+        cutoff: Maximum number of basis states to keep during reconstruction.
+
+    Returns:
+        The estimated ground-state electronic energy.
+
+    Raises:
+        ValueError: If no valid states are available after filtering the sampled counts.
+    """
+    if not counts:
+        msg = "counts must not be empty"
+        raise ValueError(msg)
+
+    basis_states = list(counts.keys())
+    basis_counts = list(counts.values())
+    num_qubits = len(basis_states[0])
+    sorted_indices = np.argsort(basis_counts)[::-1]
+
+    selected_states = []
+    for index in sorted_indices[: min(cutoff, len(basis_states)) * 2]:
+        state = basis_states[index]
+        if state.count("1") == num_qubits // 2 and (
+            state[: num_qubits // 2].count("1") == state[num_qubits // 2 :].count("1")
+        ):
+            selected_states.append(int(state, 2))
+        if len(selected_states) >= min(cutoff, len(basis_states)):
+            break
+
+    if not selected_states:
+        msg = "no states selected after occupancy filtering"
+        raise ValueError(msg)
+
+    molecule = gto.Mole(verbose=0)
+    molecule.build(atom=atom, basis=basis)
+    mean_field = scf.RHF(molecule).run()
+    core_hamiltonian = mean_field.get_hcore()
+    mean_field.kernel()
+    molecular_orbitals = mean_field.mo_coeff
+    orthogonalized_core = reduce(np.dot, (molecular_orbitals.conj().T, core_hamiltonian, molecular_orbitals))
+    orthogonalized_eri = ao2mo.kernel(mean_field._eri, molecular_orbitals)  # noqa: SLF001
+    diagonal_terms = np.kron(np.array([1, 1]), np.diag(orthogonalized_core))
+    off_diagonal_terms = orthogonalized_core - np.diag(np.diag(orthogonalized_core))
+    restored_eri = ao2mo.restore(1, orthogonalized_eri, molecule.nao)
+
+    rows, cols, values = _full_mbh_canonical_from_qsci(
+        selected_states,
+        diagonal_terms,
+        off_diagonal_terms,
+        restored_eri,
+        num_qubits,
+    )
+    reduced_hamiltonian: csr_matrix = csr_matrix(
+        (values, (rows, cols)),
+        shape=(len(selected_states), len(selected_states)),
+    )
+
+    if len(selected_states) < 3:
+        eigenvalues = np.real(sla.eigvals(reduced_hamiltonian.toarray()))
+        eigenvalues.sort()
+        return float(eigenvalues[0])
+
+    eigenvalues, _ = eigs(reduced_hamiltonian, k=1, which="SR")
+    return float(np.real(eigenvalues[0]))
 
 
 if __name__ == "__main__":
